ApiAnnotations : BooleanFormula construction discards original
Summary:
The MINIMAL pragma is captured in the parser using a BooleanFormula.

The constructors (mkBool,mkAnd,mkOr) are smart and try to minimise the boolean formula as it is constructed.

This discards the original information, making round tripping impossible.

Note: there is another version which provides a more API Annotations friendly
version of the MINIMAL pragma, but this requires changes to haddock, which will
cause problems for 7.10.2.

See https://github.com/alanz/ghc/tree/wip/10287

Test Plan: ./validate

Reviewers: hvr, austin

Subscribers: thomie, mpickering

Differential Revision: https://phabricator.haskell.org/D837

GHC Trac Issues: #10287

Conflicts:
	testsuite/tests/ghc-api/annotations/.gitignore
	testsuite/tests/ghc-api/annotations/Makefile
	testsuite/tests/ghc-api/annotations/all.T