Demand Analyzer: Do not set OneShot information (second try)
as suggested in ticket:11770#comment:1. This code was buggy
(#11770), and the occurrence analyzer does the same job anyways.

This also elaborates the notes in the occurrence analyzer accordingly.

Previously, the worker/wrapper code would go through lengths to transfer
the oneShot annotations from the original function to both the worker
and the wrapper. We now simply transfer the demand on the worker, and le
the subsequent occurrence analyzer push this onto the lambda binders.

Did not validate this yet, but I'm running out of time for today, so I’m
pushing this to a feature branch for now. A review of how I am
transfering the demand onto the worker is appreciated, maybe I’m doing
it obviously wrong.