rts: Specialize `overwritingClosure` for `MSG_THROWTO`
The function `executeMessage` replaces the `header.info` pointer with
a `WHITEHOLE` passes it to `throwToMsg` and then passes it to
`doneWithMsgThrowTo` which calls `overwritingClosure` on it. However,
`overwritingClosure` needs to know the size of the closure which is
based on the `header.info` pointer. Unfortunately, the `header.info`
pointer can't be reset to `MSG_THROWTO` or it might cause race
conditions.

The solution suggested by Simon Marlow is to add and use a function
`overwritingClosureWithSize` that takes an explicit size parameter.