rts: Fix segfault in profiling/debug/sanity checking mode
The profiling code assumes that un-used memory at the start of a block
is zero filled so the profiling code can just skip over it. However, when
compiled in DEBUG mode and with the debug sanity checking RTS flag (`-DS`)
given on the command line, the memory allocator pre-fills new blocks with
`0xaa` which causes the profiling code to segfault.

The fix is just to detect debug sanity checking and ignore any `0xaa`
filled memory. It does however mean that profiling results when debug
sanity checking is enabled may produce incorrect results.