Don't warn about name shadowing when renaming the patten in a PatSyn decl
Summary:
Previously the renamer assumed that *any* time we renamed a
pattern, the pattern was introducing new binders. This isn't true
in pattern synonym declarations where the pattern is used as part of
a definition.

We add a special case to not warn in this situation.

Test Plan:

Reviewers:

Subscribers:

GHC Trac Issues:  #12615