Introduce ConApp to Core (dead code as of yet)
This is the first step towards #12618: It adds the data constructor and
then fixes all problems due to incomplete patterns, essentially
preparing the complete compiler for the eventual use of this variant.
Because no where a ConApp is created, this should not yet have any
effect.

Care is taken to not take shortcuts via the data con worker id, as
eventually, there will be no data con worker any more.

There are a few unclear spots, marked with "TODO #12618". Input is
appreciated. These are currently:
 * CoreLint: Remove a check from the App case that will only be relevant
   occurring in the ConApp case later
 * simplExprF1: This case became very easy. I might have overlooked
   something else happening to arguments, as I read and simplified the
   code that handled App. Second pairs of eyes welcome.
 * ruleCheck: There can be no rules attached to data constructors, can
   there?
 * scExp: Can a datacon be in scSubstId?
 * dmdAnal: How to get the idStrictness equivalent of the worker?
   Or is there never something useful to be said about the strictness
   signature of an constructor? (Because strictness annotations are
   taken care of by the wrapper?