(WIP) Refactor DeriveAnyClass's instance context inference
Summary:
Currently, `DeriveAnyClass` has two glaring flaws:

* It only works on classes whose argument is of kind `*` or `* -> *` (#9821).
* The way it infers constraints makes no sense. It basically co-opts the
  algorithms used to infer contexts for `Eq` (for `*`-kinded arguments) or
  `Functor` (for `(* -> *)`-kinded arguments). This tends to produce overly
  constrained instances, which in extreme cases can lead to legitimate things
  failing to typecheck (#12594). Or even worse, it can trigger GHC panics
  (#12144 and #12423).

This completely reworks the way `DeriveAnyClass` infers constraints to fix
these two issues. It now uses the type signatures of the derived class's
methods to infer constraints (and to simplify them). A high-level description
of how this works is included in the GHC users' guide, and more technical notes
on what is going on can be found as comments (and a Note) in `TcDerivInfer`.

TODO: This patch is not yet complete. It currently uses unification to try
to make the type variables in the default and non-default type signatures
of a class's method match. This doesn't work as well as I would have imagined,
and it fails on cases that it shouldn't, causing T10361b to fail at the moment.
And in cases where it should fail, the error message is insufficient (it's
currently a GHC panic). I should find the code that gets tickled when you
manually type out an instance where the default and non-default type signatures
don't match, and emulate that.

Fixes #9821, #12144, #12423, #12594.

Test Plan: ./validate

Reviewers: simonpj, goldfire, bgamari

Subscribers:

GHC Trac Issues: #9821, #12144, #12423, #12594