Add -Wcpp-undef warning flag
Summary:
When enabled, this new warning flag passes `-Wundef` to the C
pre-processor which causes the pre-processor to warn on uses of
the `#if` directive on undefined identifiers.

It is not currently enabled in any of the standard warning groups.

Test Plan:
Add a test to this commit and make sure the test passes
on all major platforms.

Reviewers: bgamari, hvr, austin, Phyx, carter

Subscribers: thomie

Differential Revision: https://phabricator.haskell.org/D2626

GHC Trac Issues: #12752