Stop -dno-debug-output suppressing -ddump-tc-trace
The user manual states that -dno-debug-output should suppress
*unsolicited* debugging output which essentially amounts to calls
to `pprTrace`. Before I unified the interface of `traceTc` and
`traceRn`, the flag suppressed calls to `traceTc` but not to `traceRn`
or any other tracing function already controlled by a flag.

Thus, in order to make the behaviour more uniform, it seemed best to
remove this one special case.