Prevent GND from inferring an instance context for method-less classes
When `GeneralizedNewtypeDeriving` is used with a type class that has no
methods, it will generate a redundant context, and as a result, it can trigger
warnings when compiled with `-Wredundant-constraints`. This is a simple change
in behavior to check beforehand if a class has methods when deriving it with
GND, and if it has no methods, avoid inferring the redundant context.

There were a couple of testsuite changes I had to make. Notably, the test
for #3621 now fails for a completely different reason that originally intended,
as the original error arose from the inferred context. Additionally, the test
for #6088, which used to be expected to fail, now compiles without issue since
it doesn't infer a problematic instance context.

Thanks to Simon Peyton Jones for doing the necessary refactoring in
f05d685ae05ec293083f2fa7ec7ba057fbe64869.

Fixes #12814.