testsuite: Separate out Windows results for T5205
This test seems to have much different allocation behavior on Windows
and Linux.
Previously we had widened the acceptance window to 7% to accomodate
this, but
even this isn't enough any more. Instead of further widening the window
let's
just give an expected number for each platform. Really, this is
precisely the
issue with our performance testing model which I've been complaining
about
in #12758.

Fixes test for #5205 on 64-bit Windows.

Test Plan: Validate on Windows

Reviewers: austin

Subscribers: thomie, Phyx

Differential Revision: https://phabricator.haskell.org/D2848

GHC Trac Issues: #5205