Don't use $ in the definition of (<**>) in GHC.Base
Summary:
($) is special as Richard explains in the note at the top of the page.

However, when adding the note he didn't remove this usage. Normally it
didn't cause any problems as the optimiser optimised it away. However if
one had the propensity to stick one's fingers into the depths of the
inliner, it caused horrible idInfo panics.

Test Plan:

Reviewers: rwbarton

Subscribers:

GHC Trac Issues: