Always expose unfoldings for overloaded functions.
Summary:
Users expect their overloaded functions to be specialised at call sites,
however, this is only the case if they are either lucky and GHC chooses to
include the unfolding or they mark their definition with an INLINABLE pragma.
This leads to library authors marking all their functions with `INLINABLE` (or
more accurately `INLINE`) so they ensure that downstream consumers pay no cost
for their abstraction.

A more sensible default is to do this job for the library author and give more
predictable guarantees about specialisation.

Empirically, I compiled a selection of 1150 packages with (a similar) patch applied. The total size of the interface files before the patch was 519mb and after 634mb. On modern machines, I think this increase is justified for the result.

Reviewers: simonpj, austin, bgamari

Subscribers: thomie

Differential Revision: https://phabricator.haskell.org/D2929