Bump Win32 version.
Bump the version of `Win32` to `2.5.0.0` which is a major update
and includes fixes for wrong alignments and wrong 64-bit types.

The underlying type aliases have changed, so there is a potential
for user programs not to compile anymore, but the types were incorrect.

This also requires a bump in `directory`.

Test Plan: ./validate

Reviewers: bgamari, RyanGlScott, austin

Subscribers: hvr, RyanGlScott, thomie, #ghc_windows_task_force

Differential Revision: https://phabricator.haskell.org/D2938