Don't return empty initial uncovered set for an unsat context
Previously when the checker encountered an unsatisfiable term of type
context it would return an empty initial uncovered set. This caused all
pattern matches in the context to be reported as redudant.

This is arguably correct behaviour as they will never be reached but it is
better to recover and provide accurate warnings for these cases to avoid
error cascades. It would perhaps be better to report an error to the user
about an inacessible branch but this is certainly better than many confusing
redundant match warnings.

Test Plan:

Reviewers: gkaracha

Subscribers:

GHC Trac Issues: #12957