Improve the Occurrence Analyzer’s handling of one-shot functions
when determining whether an expression is used saturatedly, count the
number of value arguments that the occurrence analyser sees, and add
the number of one-shot arguments that we know (from the strictness
analyser) are passed from the context. This was suggested by Simon PJ in
ticket #13227.

perf results suggest no noticable change in allocations, reduction of
code sizes, and performance regression possibliy due to loss of join
points.

Differential Revision: https://phabricator.haskell.org/D3089