Move isJoinId, isJoinId_maybe to Id
This is just a refactoring, moving these two functions where
they belong.

The reason they were there was becuase of the use of isJoinId_maybe
in the OutputableBndr instance of TaggedBndr, which was in CoreSyn.
I moved it to PprCore, to join the OutputableBndr instance for
Var.  That makes more sense anyway.