Fix a nasty bug in CoreSubst.collectBindersPushingCo
The bug wsa in the use of (mkNthCo 0) to get the argument
part of a function coercion.  Not so! Now (->) takes four
arguments so that 0 should have been 2.

Enough with magic numbers. I defined decomposeFunCo, and used
it throughout.  Much nicer now; and correct.

The nete effect, incidentally, was that T9509 was failing to
specialise.  (And that was the initial reason for introducing
collectBindersPushingCo in the first place.)