Add -dno-debug-output to validate GhcStage1HcOpts
This flag only affects whether WARNs are printed to the build log.
ASSERT fails will still be printed and cause an abort.

Most of the WARNs in the compiler are speculative and meant to help with
debugging rather than indicative of any real errors. This causes
a lot of noise in the build log which is not helpful and makes them very
long.