Make standalone DeriveAnyClass as permissive as empty instances
Currently, GHC is quite conservative with its `deriving`-related validity
checking. In particular, it rejects several standalone anyclass-derived
instances that would otherwise be accepted if you just removed the `deriving`
part from them:

```
class C
deriving instance C -- rejected
instance C          -- accepted
```

This makes sense in a larger historical context, since before the introduction
of `DeriveAnyClass`, every derived instance was of the form:

```
instance C c1 ... cm (T t1 ... tn)
```

where `T` is a data type constructor. Many of GHC's internals surrounding
`deriving` were built up around this assumption, and when `DeriveAnyClass`
was introduced, it was tacked on it such a way that it didn't permit
instances like the example above.

This attempts to re-engineer the internals of `deriving` to rip out this
assumption. I introduce a new data type `DerivInstArgs`, which has two
possible configurations: `DataTypeInstArgs` (for 95% of instances where the
last type argument is headed by a data type constructor, as described above)
and `UnconventionalInstArgs` (for the other 5% of instances). By using this
data type diligently, we can avoid needing to refer to the "data type
constructor" everywhere (as it may not exist!).
See Note [Unconventional anyclass-derived instances] in TcDerivUtils for the
full story.

Fixes #13154.