Use actual universal tvs in check for naughty record selectors
The naughty record selector check means to limit selectors which
would
lead to existential tyvars escaping their scope. With record pattern
synonyms, there are situations where universal tyvars don't appear
in
the result type, for example:

```
pattern ReadP :: Read a => a -> String
pattern ReadP{readp} <- (read -> readp)
```

This is a similar issue to #11224 where we assumed that we can
decide
which variables are universal and which are existential by the
syntactic
check of seeing which appear in the result type. The fix is to use
"univ_tvs" from `conLikeFullSig` rather than the previous
approximation.